kwork update migration v2

+++ database/migrations/2020_06_15_080705_create_comments_table.php
@@ -15,7 +15,7 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('project_id');
+            $table->unsignedBigInteger('project_id');
             $table->string('body');
             $table->timestamps();
         });
